Use Travis for CI instead of Jenkins

+++ TestsPreprocessor.sh
@@ -1,4 +1,21 @@
-#!/usr/bin/env bash
+#!/bin/bash
+
+pwd
+
+has_param() {
+    local term="$1"
+    shift
+    for arg; do
+        if [[ $arg == "$term" ]]; then
+            return 0
+        fi
+    done
+    return 1
+}
+
+if has_param '-f' "$@"; then
+    FORCE=1;
+fi
 
 echo "# Retrieving information from pom.xml"
 readonly HDT_ARCHIVE=$(mvn -q help:evaluate -Dexpression=vibe-database.archive -DforceStdout)
@@ -7,21 +24,27 @@ readonly HPO_DOWNLOAD=$(mvn -q help:evaluate -Dexpression=hpo-owl.download -Dfor
 
 cd shared_testdata/shared
 
-echo "# Removing old data"
-rm -rf tdb # Removes deprecated TDB if still present.
-rm -rf hdt
-rm -f hp.owl
+if [ $FORCE ]; then
+  echo "# Removing old data"
+  rm -rf tdb # Removes deprecated TDB if still present.
+  rm -rf hdt
+  rm -f hp.owl
+fi
 
-echo "# Downloading data"
-curl -L -O ${HDT_DOWNLOAD} -O ${HPO_DOWNLOAD}
+if [[ ! -d hdt ]]; then
+  ##if not exists
+  echo "# Downloading data"
+  curl -L -O ${HDT_DOWNLOAD} -O ${HPO_DOWNLOAD}
 
-echo "# Preparing data for unit-tests"
-tar -xzvf ${HDT_ARCHIVE}
-mv ${HDT_ARCHIVE%.tar.gz} hdt
-cd hdt
-mv ${HDT_ARCHIVE%-hdt.tar.gz}.hdt vibe.hdt
-mv ${HDT_ARCHIVE%-hdt.tar.gz}.hdt.index.v1-1 vibe.hdt.index.v1-1
-cd ../
-rm ${HDT_ARCHIVE}
-
+  echo "# Preparing data for unit-tests"
+  tar -xzvf ${HDT_ARCHIVE}
+  mv ${HDT_ARCHIVE%.tar.gz} hdt
+  cd hdt
+  mv ${HDT_ARCHIVE%-hdt.tar.gz}.hdt vibe.hdt
+  mv ${HDT_ARCHIVE%-hdt.tar.gz}.hdt.index.v1-1 vibe.hdt.index.v1-1
+  cd ../
+  rm ${HDT_ARCHIVE}
+else
+  echo "Data already exists, skipping download."
+fi
 cd ../../
